Refactoring components
Added components folder
Added components files within folder
