+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,9 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\public",
+    "\\src"
+  ],
+  "SelectedNode": "\\public\\index.html",
+  "PreviewInSolutionExplorer": false
+}
